Implement upgrade system and refactor player stats
- Add new upgrade types (Move Speed, Projectile Speed, Pickup Radius, XP Boost) and create a centralized PlayerStats component that manages all player attributes.
- Upgrades now apply modifiers through PlayerStats instead of direct script modifications.
- Update gameplay systems to use PlayerStats properties for movement, damage, fire rate, and XP rewards.
- Change XP system from int to float for finer granularity.
- Fix UpgradePanelUI boundary check and Fire Rate Upgrade multiplier.

+++ Assets/Scripts/Data/LevelDatabase.cs
@@ -14,5 +14,5 @@ public class LevelDatabase : ScriptableObject
 public class LevelData
 {
     public int level;
-    public int xpRequired;
+    public float xpRequired;
 }
