+++ Assets/Scripts/Gameplay/XP/PlayerExperience.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 
 public class PlayerExperience : MonoBehaviour
 {
-    public event Action<int, int> OnXPChanged;
+    public event Action<float, float> OnXPChanged;
     [SerializeField] private LevelDatabase levelDatabase;
     [SerializeField] private UpgradeManager upgradeManager;
     private int currentLevel = 1;
-    private int currentXP;
-    private int XPToNextLevel => levelDatabase.GetLevelData(currentLevel).xpRequired;
-    public void AddXP(int amount)
+    private float currentXP;
+    private float XPToNextLevel => levelDatabase.GetLevelData(currentLevel).xpRequired;
+    public void AddXP(float amount)
     {
         currentXP += amount;
         while(currentXP >= XPToNextLevel)
